requirements.txt
anthropic
python-docx==1.2.0
pypdf2==3.0.1
